// app/src/app/core/FitActivityEnum.enum.ts
// exported from https://developers.google.com/fit/rest/v1/reference/activity-types
export enum FitActivityType {
    AEROBICS = 9,
    ARCHERY = 119,
    BADMINTON = 10,
    BASEBALL = 11,
    BASKETBALL = 12,
    BIATHLON = 13,
    BIKING = 1,
    HANDBIKING = 14,
    MOUNTAIN_BIKING = 15,
    ROAD_BIKING = 16,
    SPINNING = 17,
    STATIONARY_BIKING = 18,
    UTILITY_BIKING = 19,
    BOXING = 20,
    CALISTHENICS = 21,
    CIRCUIT_TRAINING = 22,
    CRICKET = 23,
    CROSSFIT = 113,
    CURLING = 106,
    DANCING = 24,
    DIVING = 102,
    ELEVATOR = 117,
    ELLIPTICAL = 25,
    ERGOMETER = 103,
    ESCALATOR = 118,
    FENCING = 26,
    FOOTBALL_AMERICAN = 27,
    FOOTBALL_AUSTRALIAN = 28,
    FOOTBALL_SOCCER = 29,
    FRISBEE = 30,
    GARDENING = 31,
    GOLF = 32,
    GYMNASTICS = 33,
    HANDBALL = 34,
    HIIT = 114,
    HIKING = 35,
    HOCKEY = 36,
    HORSEBACK_RIDING = 37,
    HOUSEWORK = 38,
    ICE_SKATING = 104,
    IN_VEHICLE = 0,
    INTERVAL_TRAINING = 115,
    JUMPING_ROPE = 39,
    KAYAKING = 40,
    KETTLEBELL_TRAINING = 41,
    KICKBOXING = 42,
    KITESURFING = 43,
    MARTIAL_ARTS = 44,
    MEDITATION = 45,
    MIXED_MARTIAL_ARTS = 46,
    ON_FOOT = 2,
    OTHER_UNCLASSIFIED_FITNESS_ACTIVITY = 108,
    P90X_EXERCISES = 47,
    PARAGLIDING = 48,
    PILATES = 49,
    POLO = 50,
    RACQUETBALL = 51,
    ROCK_CLIMBING = 52,
    ROWING = 53,
    ROWING_MACHINE = 54,
    RUGBY = 55,
    RUNNING = 8,
    JOGGING = 56,
    RUNNING_ON_SAND = 57,
    RUNNING_TREADMILL = 58,
    SAILING = 59,
    SCUBA_DIVING = 60,
    SKATEBOARDING = 61,
    SKATING = 62,
    CROSS_SKATING = 63,
    INDOOR_SKATING = 105,
    INLINE_SKATING_ROLLERBLADING = 64,
    SKIING = 65,
    BACK_COUNTRY_SKIING = 66,
    CROSS_COUNTRY_SKIING = 67,
    DOWNHILL_SKIING = 68,
    KITE_SKIING = 69,
    ROLLER_SKIING = 70,
    SLEDDING = 71,
    SLEEPING = 72,
    LIGHT_SLEEP = 109, // relevant sleep type
    DEEP_SLEEP = 110, // relevant sleep type
    REM_SLEEP = 111, // relevant sleep type
    AWAKE_DURING_SLEEP_CYCLE = 112, // relevant sleep type
    SNOWBOARDING = 73,
    SNOWMOBILE = 74,
    SNOWSHOEING = 75,
    SOFTBALL = 120,
    SQUASH = 76,
    STAIR_CLIMBING = 77,
    STAIR_CLIMBING_MACHINE78,
    STAND_UP_PADDLEBOARDING79,
    STILL_NOT_MOVING = 3,
    STRENGTH_TRAINING = 80,
    SURFING = 81,
    SWIMMING = 82,
    SWIMMING_OPEN_WATER = 84,
    SWIMMING_SWIMMING_POOL = 83,
    TABLE_TENNIS_PING_PONG85,
    TEAM_SPORTS = 86,
    TENNIS = 87,
    TILTING_SUDDEN_DEVICE_GRAVITY_CHANGE = 5,
    TREADMILL_WALKING_OR_RUNNING88,
    UNKNOWN_UNABLE_TO_DETECT_ACTIVITY = 4,
    VOLLEYBALL = 89,
    VOLLEYBALL_BEACH = 90,
    VOLLEYBALL_INDOOR = 91,
    WAKEBOARDING = 92,
    WALKING = 7,
    WALKING_FITNESS = 93,
    NORDING_WALKING = 94,
    WALKING_TREADMILL = 95,
    WALKING_STROLLER = 116,
    WATERPOLO = 96,
    WEIGHTLIFTING = 97,
    WHEELCHAIR = 98,
    WINDSURFING = 99,
    YOGA = 100,
    ZUMBA = 101,
}